essential-init, cube-essential: Fix for read-only files in /etc and /var
This commit cleans up the remaining configuration files in /etc and
/var which are needed to boot the system with a fully read-only root
file system.

The script which handles the persistence was moved to
essential-opt-mount where previously it was a single command to
execute "mount /opt/container".  All of these commands in the script
need to run to setup any bind mounts to the persistent area before
trying to start services such as systemd-networkd.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>